Refactor CategoryFilter component
